requirements.txt
streamlit
pandas
pandas-ta
requests
python-dotenv
openai
plotly
jwt
pytz